backend/requirements.txt
 Django==5.0.6
 djangorestframework==3.15.2
 django-allauth==0.63.6
 dj-rest-auth==5.0.2
 djangorestframework-simplejwt==5.3.1
 django-cors-headers==4.4.0
 psycopg2-binary==2.9.9
 python-dotenv==1.0.1
 celery==5.4.0
 redis==5.0.4
 pdfplumber==0.11.4
 python-docx==1.1.2
 requests==2.32.3
 PyJWT==2.9.0
 fpdf2==2.7.9
